Better error handling with messages

+++ bdimensions.py
@@ -21,16 +21,19 @@ def hello_world():
 def users():
     if request.method == 'POST': # Add a new user
         if request.headers['Content-Type'] == 'application/json':
-            d = request.get_json()
-            return addNewUser(d['firstName'],d['lastName'],d['genre'],d['dateOfBirth'])
+            if request.get_json(silent=True):
+                d = request.get_json()
+                return addNewUser(d)
+            else:
+                abort (400, 'No json in request')
         else:
-            abort(415)
+            abort(415, 'Content Type must be application/json')
     elif request.method == 'GET': # List all users
        return getAllUsers()
     elif request.method == 'PUT':
-        return 'Bulk update users'
+        abort(501, 'Bulk update users not imlemented')
     elif request.method == 'DELETE':
-        return 'Delete all users'
+        abort(501, 'Delete all users not imlemented')
 
 
 @app.route('/users/<int:user_id>', methods = ['GET', 'PUT', 'DELETE'])
@@ -39,13 +42,15 @@ def user(user_id):
         return getUser(user_id)
     elif request.method == 'PUT':
         if request.headers['Content-Type'] == 'application/json':
-            d = request.get_json()
-            return updateUser(user_id, d)
+            if request.get_json(silent=True):
+                d = request.get_json()
+                return updateUser(user_id, d)
+            else:
+                abort (400, 'No json in request')
         else:
-            abort(415)
-        #return 'If exists update user with id %s' % user_id
+            abort(415, 'Content Type must be application/json')
     elif request.method == 'DELETE':
-        return 'Delete user with id %s' % user_id
+        return deleteUser(user_id)
 
 
 @app.route('/users/<int:user_id>/data', methods = ['POST', 'GET', 'DELETE'])
@@ -68,40 +73,67 @@ def data(user_id, data_id):
         return 'Delete data item with id %s for user with id %s' % (data_id, user_id)
 
 
-def addNewUser(firstName,lastName,genre,dateOfBirth):
-    user = Users(firstName = firstName, lastName = lastName, genre = genre, dateOfBirth = dateOfBirth)
-    session.add(user)
-    session.commit()
-    return jsonify(Users=user.serialize)
+def addNewUser(d):
+    if 'firstName' in d and d['firstName'] \
+    and 'lastName' in d and d['lastName'] \
+    and 'genre' in d and d['genre'] \
+    and 'dateOfBirth' in d and d['dateOfBirth']:
+        user = Users( \
+        firstName = d['firstName'], \
+        lastName = d['lastName'], \
+        genre = d['lastName'], \
+        dateOfBirth = d['dateOfBirth'])
+        session.add(user)
+        session.commit()
+        return jsonify(Users=user.serialize)
+    else:
+        abort(400, 'Request must contain keywords firstName, lastName, genre and dateOfBirth with values')
 
 def getAllUsers():
     users = session.query(Users).all()
     return jsonify(Users=[i.serialize for i in users])
 
 def getUser(user_id):
-    user = session.query(Users).filter_by(id = user_id).one()
-    return jsonify(Users=user.serialize)
+    user = session.query(Users).filter_by(id = user_id).first()
+    if user:
+        return jsonify(Users=user.serialize)
+    else:
+        abort(404, 'User with id %s does not exist' % user_id)
 
 def updateUser(user_id, d):
-    user = session.query(Users).filter_by(id = user_id).one()
-    if 'firstName' in d:
-        user.firstName = d['firstName']
-    if 'lastName' in d:
-        user.lastName = d['lastName']
-    if 'genre' in d:
-        user.genre = d['genre']
-    if 'dateOfBirth' in d:
-        user.dateOfBirth = d['dateOfBirth']
-    session.add(user)
-    session.commit()
-    return jsonify(Users=user.serialize)
+    user = session.query(Users).filter_by(id = user_id).first()
+    if user:
+        i = 0
+        if 'firstName' in d and d['firstName']:
+            i += 1
+            user.firstName = d['firstName']
+        if 'lastName' in d and d['lastName']:
+            i += 1
+            user.lastName = d['lastName']
+        if 'genre' in d and d['genre']:
+            i += 1
+            user.genre = d['genre']
+        if 'dateOfBirth' in d and d['dateOfBirth']:
+            i += 1
+            user.dateOfBirth = d['dateOfBirth']
+        if (i > 0):
+            session.add(user)
+            session.commit()
+            return jsonify(Users=user.serialize)
+        else:
+            abort(400, 'Request must contain firstName, lastName, genre or dateOfBirth keyword with values')
+    else:
+        abort(404, 'User with id %s does not exist' % user_id)
 
 
 def deleteUser(user_id):
-    user = session.query(Users).filter_by(id = user_id).one()
-    session.delete(user)
-    session.commit()
-    return 'Removed user with id %s' % user_id
+    user = session.query(Users).filter_by(id = user_id).first()
+    if user:
+        session.delete(user)
+        session.commit()
+        return 'Removed user with id %s' % user_id
+    else:
+        abort(404, 'User with id %s does not exist' % user_id)
 
 
 if __name__ == '__main__':
